+++ key-annotations/_utils/normalized-name.ts
@@ -1,0 +1,17 @@
+import { SpecialKeys, SPECIAL_KEY_PREFIX_TOKEN } from "../keys";
+import { specialKeyTypeFrom } from "./annotation-from-name";
+
+/**
+ * case 1. //@ignore/some-name : null
+ * case 2. //@data-target/some-name : "some-name"
+ * case 3. Rectangle 1 : "Rectangle 1"
+ * @param name
+ * @returns
+ */
+export function normalizedName(name: string) {
+  if (specialKeyTypeFrom(name) == SpecialKeys.KEY_IGNORE_ALL) {
+    return null;
+  }
+
+  return name.replace(SPECIAL_KEY_PREFIX_TOKEN, "");
+}
